+++ tests/unittests/test_dev_mode.py
@@ -2,6 +2,7 @@ from asyncio import wait_for
 import unittest
 import os
 import json
+import tempfile
 import requests
 from tap_deputy.client import DeputyClient, Server401TokenExpiredError
 from unittest import mock
@@ -18,7 +19,7 @@ test_config = {
     "refresh_token": "old_refresh_token",
     "access_token": "old_access_token"
 }
-test_config_path = "/tmp/test_config.json"
+test_config_path = os.path.join(tempfile.gettempdir(), "test_config.json")
 
 class Mockresponse:
     """ Mock response object class."""
@@ -106,7 +107,7 @@ class TestDevMode(unittest.TestCase):
 
         self.assertEqual(deputy.refresh_token, "old_refresh_token")
         self.assertEqual(deputy.access_token, "old_access_token")
-        self.assertEquals(mocked_refresh.call_count, 0)
+        self.assertEqual(mocked_refresh.call_count, 0)
 
     @mock.patch('tap_deputy.client.DeputyClient.post')
     def test_dev_mode_enabled_valid_token(self, mocked_post_request):
@@ -135,4 +136,4 @@ class TestDevMode(unittest.TestCase):
                            auth_call=True)
 
         # Verify the call count for each error
-        self.assertEquals(mocked_post_request.call_count, retry_count)
+        self.assertEqual(mocked_post_request.call_count, retry_count)